# output/final_deliverables/D-20_fastapi/openapi.json
{"openapi":"3.1.0","info":{"title":"NIFTY 100 Financial Intelligence Platform API","description":"REST API for the NIFTY 100 Financial Intelligence Platform. Provides company, screening, sector, peer, valuation, portfolio, document and health endpoints.","version":"1.0.0"},"paths":{"/":{"get":{"summary":"Root","description":"API root endpoint.","operationId":"root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1":{"get":{"summary":"Api Info","description":"Return API information.","operationId":"api_info_api_v1_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/companies/":{"get":{"tags":["Companies"],"summary":"Get Companies","description":"Return all NIFTY 100 companies.\n\nIncludes:\n- company ID\n- company name\n- sector\n- industry","operationId":"get_companies_api_v1_companies__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/companies/{company_id}":{"get":{"tags":["Companies"],"summary":"Get Company","description":"Return complete company master information.","operationId":"get_company_api_v1_companies__company_id__get","parameters":[{"name":"company_id","in":"path","required":true,"schema":{"type":"string","title":"Company Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{company_id}/financials":{"get":{"tags":["Companies"],"summary":"Get Company Financials","description":"Return historical profit and loss information.","operationId":"get_company_financials_api_v1_companies__company_id__financials_get","parameters":[{"name":"company_id","in":"path","required":true,"schema":{"type":"string","title":"Company Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{company_id}/ratios":{"get":{"tags":["Companies"],"summary":"Get Company Ratios","description":"Return historical financial ratios and KPI calculations.","operationId":"get_company_ratios_api_v1_companies__company_id__ratios_get","parameters":[{"name":"company_id","in":"path","required":true,"schema":{"type":"string","title":"Company Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{company_id}/valuation":{"get":{"tags":["Companies"],"summary":"Get Company Valuation","description":"Return historical market-cap and valuation metrics.","operationId":"get_company_valuation_api_v1_companies__company_id__valuation_get","parameters":[{"name":"company_id","in":"path","required":true,"schema":{"type":"string","title":"Company Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{company_id}/peers":{"get":{"tags":["Companies"],"summary":"Get Company Peers","description":"Return companies belonging to the same peer group.\n\nA peer group is defined by the peer_groups table.\nThe requested company itself is excluded from the result.\n\nIf the company has no peer-group assignment, an empty\npeer list is returned rather than inventing a group.","operationId":"get_company_peers_api_v1_companies__company_id__peers_get","parameters":[{"name":"company_id","in":"path","required":true,"schema":{"type":"string","title":"Company Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/companies/{company_id}/documents":{"get":{"tags":["Companies"],"summary":"Get Company Documents","description":"Return documents associated with a company.","operationId":"get_company_documents_api_v1_companies__company_id__documents_get","parameters":[{"name":"company_id","in":"path","required":true,"schema":{"type":"string","title":"Company Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/screener/":{"get":{"tags":["Screener"],"summary":"Screen Companies","description":"Screen companies using latest available financial-ratio data.\n\nOptional filters:\n- sector\n- min_roe\n- max_debt_to_equity\n- min_revenue_cagr_5yr\n- min_opm\n- min_fcf","operationId":"screen_companies_api_v1_screener__get","parameters":[{"name":"sector","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sector"}},{"name":"min_roe","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Min Roe"}},{"name":"max_debt_to_equity","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Max Debt To Equity"}},{"name":"min_revenue_cagr_5yr","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Min Revenue Cagr 5Yr"}},{"name":"min_opm","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Min Opm"}},{"name":"min_fcf","in":"query","required":false,"schema":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Min Fcf"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/sectors/":{"get":{"tags":["Sectors"],"summary":"List Sectors","description":"Return all sectors with company counts.","operationId":"list_sectors_api_v1_sectors__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/sectors/{sector_name}":{"get":{"tags":["Sectors"],"summary":"Get Sector","description":"Return companies belonging to a sector.","operationId":"get_sector_api_v1_sectors__sector_name__get","parameters":[{"name":"sector_name","in":"path","required":true,"schema":{"type":"string","title":"Sector Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/sectors/{sector_name}/summary":{"get":{"tags":["Sectors"],"summary":"Sector Summary","description":"Return latest financial-ratio summary for a sector.","operationId":"sector_summary_api_v1_sectors__sector_name__summary_get","parameters":[{"name":"sector_name","in":"path","required":true,"schema":{"type":"string","title":"Sector Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/peers/{company_id}":{"get":{"tags":["Peers"],"summary":"Peer Comparison","description":"Return peer-group comparison for a company using peer_percentiles.","operationId":"peer_comparison_api_v1_peers__company_id__get","parameters":[{"name":"company_id","in":"path","required":true,"schema":{"type":"string","title":"Company Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/valuation/market-cap":{"get":{"tags":["Valuation"],"summary":"Market Cap","operationId":"market_cap_api_v1_valuation_market_cap_get","parameters":[{"name":"company_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Company Id"}},{"name":"year","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Year"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/valuation/market-cap/{company_id}":{"get":{"tags":["Valuation"],"summary":"Company Market Cap","operationId":"company_market_cap_api_v1_valuation_market_cap__company_id__get","parameters":[{"name":"company_id","in":"path","required":true,"schema":{"type":"string","title":"Company Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/portfolio/stats":{"get":{"tags":["Portfolio"],"summary":"Portfolio Statistics","operationId":"portfolio_statistics_api_v1_portfolio_stats_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/documents/":{"get":{"tags":["Documents"],"summary":"Search Documents","operationId":"search_documents_api_v1_documents__get","parameters":[{"name":"company_id","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Company Id"}},{"name":"year","in":"query","required":false,"schema":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Year"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/health/":{"get":{"tags":["Health"],"summary":"Health Check","description":"Basic API and database health check.","operationId":"health_check_api_v1_health__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/system/health":{"get":{"summary":"System Health","description":"Detailed system health information.\n\nIncludes:\n- API status\n- database status\n- table row counts\n- uptime\n- version","operationId":"system_health_api_v1_system_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}